apps/organisations: fix language chooser in organisation form - fixes #1768
It makes the buttons switch between the languages, and all languages
addable (if they aren't already there) and removable (if they are there)
on the plus-button.
The jQuery was there to choose the right language for the tabs, which
are in the html. So, currently if a language is removed that
was active on the tabs, The tab content is still shown, while no button
is active and the respective language is removed when saving.